Codecs/FLAC: Fix 32-bit truncation of large bitstream seeks
setBitPosition(uint64_t) narrowed its argument to uint32_t before handing it to
skipBits(), so seeking past 2^32 bits (~512 MB into the bitstream) silently
truncated to the low 32 bits and landed at the wrong position. Widen skipBits()
to take uint64_t (it already advances in 32-bit chunks, so it handles any count)
and pass the position through unmodified.

+++ include/codecs/flac/BitstreamReader.h
@@ -48,7 +48,7 @@ public:
     // Alignment
     bool alignToByte();
     bool isAligned() const;
-    bool skipBits(uint32_t bit_count);
+    bool skipBits(uint64_t bit_count);
     
     // Position tracking
     uint64_t getBitPosition() const;
